requirements.txt
streamlit
python-dotenv
pyttsx3
google-generativeai
pip install gTTS
